Semana 9

+++ Keep_Health_Project/src/app/pages/cadastro/cadastro.component.ts
@@ -7,27 +7,40 @@ import { Router } from '@angular/router';
   styleUrls: ['./cadastro.component.css']
 })
 export class CadastroComponent {
+  name: string | undefined;
   email: string | undefined;
+  birthdate: string | undefined;
   password: string | undefined;
+  confirmPassword: string | undefined;
 
   constructor(private router: Router) {}
 
   register() {
+    if (this.password !== this.confirmPassword) {
+      alert('As senhas não coincidem');
+      return;
+    }
+
     let users = localStorage.getItem('users');
     let parsedUsers = users ? JSON.parse(users) : [];
 
-    let userExists = parsedUsers.some((user: { email: string | undefined; }) => user.email === this.email);
+    let userExists = parsedUsers.some((user: { email: string; }) => user.email === this.email);
 
     if (userExists) {
       // Exibe um alerta se o email já estiver cadastrado
       alert('Email já cadastrado');
     } else {
       // Adiciona o novo usuário à lista de usuários e salva no localStorage
-      parsedUsers.push({ email: this.email, password: this.password });
+      parsedUsers.push({ name: this.name, email: this.email, birthdate: this.birthdate, password: this.password });
       localStorage.setItem('users', JSON.stringify(parsedUsers));
 
       // Redireciona para a home após o cadastro
       this.router.navigate(['/home']);
     }
+  }
+
+  goBack() {
+    // Redireciona para a tela de login
+    this.router.navigate(['/login']);
   }
 }
